+++ mail_2_rmq.py
@@ -469,7 +469,7 @@ def process_message(cfg, log):
     log.log_info("Parsing email...")
     msg = parse_email()
     subj = gen_libs.pascalize(filter_subject(msg["subject"], cfg))
-    email_list = get_email_addr(msg["from"])
+    email_list = gen_libs.find_email_addr(msg["from"])
     from_addr = email_list[0] if email_list else None
     log.log_info("Instance creation")